% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema_parsing.R
\name{all_enums_per_col}
\alias{all_enums_per_col}
\title{Get all the enums for each slot}
\usage{
all_enums_per_col(schema)
}
\arguments{
\item{schema}{A schema file returned from \link{load_schema}}
}
\value{
A named list of the slots and their enums
}
\description{
Return a giant list of the slots and their corresponding enum entries
}
\keyword{internal,}
\keyword{parsing}
