% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema_parsing.R
\name{categories}
\alias{categories}
\title{Get the categories that the schema uses to group its slots}
\usage{
categories(schema)
}
\arguments{
\item{schema}{A schema file returned from \link{load_schema}}
}
\value{
a vector of the categories of the schema
}
\description{
Get the categories that the schema uses to group its slots
}
\keyword{internal,}
\keyword{parsing}
