% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema_parsing.R
\name{check_multivalues}
\alias{check_multivalues}
\title{Check if the slots are multivalued}
\usage{
check_multivalues(schema, slots)
}
\arguments{
\item{schema}{A schema file returned from \link{load_schema}}

\item{slots}{A vector of slots}
}
\value{
Logical vector, TRUE if the list is multivalued, FALSE if not.
}
\description{
Check if the slots are multivalued
}
