% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{check_slots_in_schema}
\alias{check_slots_in_schema}
\title{Ensure that the slot is in the schema!}
\usage{
check_slots_in_schema(schema, slots)
}
\description{
Ensure that the slot is in the schema!
}
\keyword{internal,errors}
