% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema_parsing.R
\name{convert_range_token_to_date}
\alias{convert_range_token_to_date}
\title{Convert the min or max range tokens to actual dates to check against}
\usage{
convert_range_token_to_date(token, slot, dataframe)
}
\arguments{
\item{token}{A string, taken from the "todo" value of a date slot.}

\item{slot}{The slot name, used for}

\item{dataframe}{The full dataframe of the data, we need this to extract the dates in the event that
the min value is acutally another column.}
}
\value{
description
}
\description{
Convert the min or max range tokens to actual dates to check against
}
\keyword{internal,}
\keyword{parsing}
