% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema_parsing.R
\name{download_schema}
\alias{download_schema}
\title{Download a schema file from a url}
\usage{
download_schema(yaml_url)
}
\description{
Download a schema file from a url
}
\keyword{internal,}
\keyword{parsing}
