% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema_parsing.R
\name{get_all_field_ontology_terms}
\alias{get_all_field_ontology_terms}
\title{Get a list of all the ontology terms used in the standard.}
\usage{
get_all_field_ontology_terms(schema)
}
\arguments{
\item{schema}{A schema file returned from \link{load_schema}}
}
\value{
A named vector of all ontology terms from columns with menus, where
the name of each value is the specification field it belongs to.
}
\description{
This function retrieves a list of all the enums (terms) used in a standard,
returned as a named list. Organisational terms and nulls are removed from
this list, for ease of searching through them
}
