% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema_parsing.R
\name{get_category}
\alias{get_category}
\title{Get the category of the slots}
\usage{
get_category(schema, slots)
}
\arguments{
\item{schema}{A schema file returned from \link{load_schema}}

\item{slots}{A vector of slots}
}
\value{
A named list of the slots with their category
}
\description{
Get the category of the slots
}
