% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema_parsing.R
\name{get_date_range_as_interval}
\alias{get_date_range_as_interval}
\title{Get the date range of a slot}
\usage{
get_date_range_as_interval(schema, slot)
}
\arguments{
\item{schema}{A schema file returned from \link{load_schema}}

\item{slot}{A name of a single slot}
}
\value{
A \link[lubridate:interval]{lubridate::interval} of the acceptable date range.
}
\description{
This returns the acceptable date range of a date slot.
Slot date ranges appear to be coded in a strange 'todo' section of the slot.
Parse this section (will likely error if not present) into a lubridate interval.
A '{today}' value is parsed as the current date.
}
\keyword{internal,}
\keyword{parsing}
