% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema_parsing.R
\name{get_field_importance}
\alias{get_field_importance}
\title{Get the importance of a slot}
\usage{
get_field_importance(schema, slot)
}
\arguments{
\item{schema}{A schema file returned from \link{load_schema}}

\item{slot}{A name of a single slot}
}
\value{
The importance of the field
}
\description{
The standards often define the importance of a slot.
Retrieve this for a single slot.
}
\keyword{internal}
