% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema_parsing.R
\name{get_first_identification_col}
\alias{get_first_identification_col}
\title{Get the identification column, returning the 1st one in there are multiple.}
\usage{
get_first_identification_col(schema)
}
\arguments{
\item{schema}{A schema file returned from \link{load_schema}}
}
\description{
Get the identification column, returning the 1st one in there are multiple.
}
