% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema_parsing.R
\name{get_info}
\alias{get_info}
\title{Return the description and comments tags of a GRDI field}
\usage{
get_info(schema, slot)
}
\arguments{
\item{schema}{A schema file returned from \link{load_schema}}
}
\description{
Return the description and comments tags of a GRDI field
}
