% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema_parsing.R
\name{get_menu_values}
\alias{get_menu_values}
\title{Get the permissible values of a slot}
\usage{
get_menu_values(schema, slot)
}
\arguments{
\item{schema}{A schema file returned from \link{load_schema}}

\item{slot}{A name of a single slot}
}
\value{
A vector of permissible values for the slot
}
\description{
Some slots are restricted to a picklist.
Get the permissible values of the slot
}
\keyword{internal}
