% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{get_missing_schema_cols}
\alias{get_missing_schema_cols}
\title{Test for missing schema columns in the data}
\usage{
get_missing_schema_cols(schema, data)
}
\arguments{
\item{schema}{The schema}

\item{data}{Dataframe to test}
}
\value{
Nothing, but prints a log of columns that are missing in the dataframe
}
\description{
Test for missing schema columns in the data
}
\keyword{internal,}
\keyword{validation}
