% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema_parsing.R
\name{get_null_value}
\alias{get_null_value}
\title{Get the null value of the schema}
\usage{
get_null_value(schema, x = NULL)
}
\arguments{
\item{schema}{A schema file returned from \link{load_schema}}

\item{x}{Qery to search for a specific ontology term}
}
\value{
A schema compliant null value
}
\description{
Returns a valid NULL value. If no query is supplied to search for a specific
ontology term, than a list of all the possible NULL values are supplied.
}
\keyword{internal}
