% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema_parsing.R
\name{get_pattern}
\alias{get_pattern}
\title{Get the regex pattern of the slots}
\usage{
get_pattern(schema, slots)
}
\arguments{
\item{schema}{A schema file returned from \link{load_schema}}

\item{slots}{A vector of slots}
}
\value{
a named vector of the regex that applies to the slot, or NA if none present
}
\description{
Get the regex pattern of the slots
}
\keyword{internal,}
\keyword{parsing}
