% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema_parsing.R
\name{get_permissible_values}
\alias{get_permissible_values}
\title{Get all the permissible values of a slot, if they exist.}
\usage{
get_permissible_values(schema, slot)
}
\arguments{
\item{schema}{A schema file returned from \link{load_schema}}

\item{slot}{A name of a single slot}
}
\value{
A vector of permissible values for the slot
}
\description{
This gets the permissible values of each range and appends them into
a single vector. This includes the values of the NullValueMenu
}
\keyword{internal,}
\keyword{parsing}
