% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema_parsing.R
\name{get_slot_type}
\alias{get_slot_type}
\title{Get the type of the slot}
\usage{
get_slot_type(schema, slot)
}
\arguments{
\item{schema}{A schema file returned from \link{load_schema}}
}
\value{
The type of the slot
}
\description{
Get the type of the slot, inferred from the 'range' value when 'Menu' items are removed.
If all the 'Menu' items are removed, it must be a menu type so return that instead.
}
\keyword{internal,}
\keyword{parsing}
