% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema_parsing.R
\name{get_slots_per_cat}
\alias{get_slots_per_cat}
\title{Get all the slots that belong to the supplied category}
\usage{
get_slots_per_cat(schema, category)
}
\arguments{
\item{schema}{A schema file returned from \link{load_schema}}

\item{category}{A single category}
}
\value{
a vector of the slots that belong to the category
}
\description{
Get all the slots that belong to the supplied category
}
\keyword{internal,}
\keyword{parsing}
