% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_and_replace.R
\name{grep_field_val}
\alias{grep_field_val}
\title{Search for a valid ontology term from a GRDI-AMR2 field}
\usage{
grep_field_val(schema, field, x, ...)
}
\arguments{
\item{field}{The GRDI field to be searched over}

\item{x}{Search term to be passed onto agrep}

\item{...}{Other arguments passed onto \code{\link[=agrep]{agrep()}}}
}
\description{
Search for a valid ontology term from a GRDI-AMR2 field
}
