% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{has_whitespace}
\alias{has_whitespace}
\title{Check for whitespace}
\usage{
has_whitespace(data)
}
\arguments{
\item{data}{vector of values to check}
}
\value{
logical vector, TRUE if no whitespace, FALSE if dectected
}
\description{
Check for whitespace
}
\keyword{internal,}
\keyword{validation}
