% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{is_date}
\alias{is_date}
\title{Are the values expressed in Year-Month-Date format?}
\usage{
is_date(x)
}
\arguments{
\item{x}{A vector of strings to test.}
}
\value{
Logical vector indicating if the string could be coerced to date format
}
\description{
Are the values expressed in Year-Month-Date format?
}
\keyword{internal,}
\keyword{validation}
