% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{is_decimal}
\alias{is_decimal}
\title{Are the values decimals?}
\usage{
is_decimal(x)
}
\arguments{
\item{x}{A vextor of strings to test.}
}
\value{
Logical vector indicating if the string could be coerced to decimal format
}
\description{
Are the values decimals?
}
\keyword{internal,}
\keyword{validation}
