% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema_parsing.R
\name{is_identifier}
\alias{is_identifier}
\title{Determine if the slots are identifier columns.}
\usage{
is_identifier(schema, slots)
}
\arguments{
\item{schema}{A schema file returned from \link{load_schema}}

\item{slots}{A vector of slots}
}
\value{
Logical vector, TRUE if idendifier, FALSE if not.
}
\description{
Determine if the slots are identifier columns.
}
\keyword{internal,}
\keyword{parsing}
