% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{is_permissible_value}
\alias{is_permissible_value}
\title{Return TRUE if x matches val OR is NA!}
\usage{
is_permissible_value(x, vals)
}
\arguments{
\item{x}{a vector}

\item{vals}{values to match against}
}
\value{
Logical vector
}
\description{
Converts both input values and permissible values to lowercase
}
