% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema_parsing.R
\name{load_schema}
\alias{load_schema}
\title{Load a schema file}
\usage{
load_schema(file)
}
\arguments{
\item{file}{The path to the yaml schema file}
}
\value{
A schema file -> really just a parsed YAML loaded into R.
}
\description{
Load a schema file
}
\keyword{internal,}
\keyword{parsing}
