% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{log_basic_validation}
\alias{log_basic_validation}
\title{Print out if validation passed or failed}
\usage{
log_basic_validation(x, slot, type, data, ...)
}
\arguments{
\item{x}{Logical vector indication pass/fail of validation}

\item{type}{The 'type' of the slot, returned from \link{get_slot_type}}

\item{data}{the vector of data that was validated}
}
\value{
Nothing, but prints out if validation succeeded or failed.
}
\description{
Print out if validation passed or failed
}
\keyword{internal,}
\keyword{validation}
