% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{log_error}
\alias{log_error}
\title{Simply log an error and call the slot name!}
\usage{
log_error(slot, ...)
}
\arguments{
\item{slot}{the name of the slot!}
}
\description{
Simply log an error and call the slot name!
}
