% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{log_message}
\alias{log_message}
\title{General logging function}
\usage{
log_message(type = c("error", "warning", "pass", "note"), slot = NULL, ...)
}
\arguments{
\item{type}{One of either error, warning, pass, or note.}

\item{slot}{the name of the slot, prepended to the log message if not null}

\item{...}{the rest of the message to print}
}
\value{
Nothing, but will print the log using cat.
}
\description{
The logging function, which uses cat and crayon, along with the option DHRtools.loglevel to
print out errors, warnings, passes, and notes, depending on the desired log level.
}
\keyword{internal}
\keyword{logging,}
