% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{log_warning}
\alias{log_warning}
\title{Log a warning}
\usage{
log_warning(slot, ...)
}
\arguments{
\item{slot}{the name of the slot!}

\item{...}{Additional items to print via cat}
}
\description{
Log a warning
}
