% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema_parsing.R
\name{min_max_value}
\alias{min_max_value}
\title{Get the minimum and maximum value of a slot}
\usage{
min_max_value(schema, slot)
}
\arguments{
\item{schema}{A schema file returned from \link{load_schema}}

\item{slot}{A name of a single slot}
}
\value{
vector of length 2 (min,max), or NA if there is no range
}
\description{
Get the minimum and maximum value of a slot
}
