% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{parse_date}
\alias{parse_date}
\title{The date parsing this package will use.}
\usage{
parse_date(x)
}
\arguments{
\item{x}{A vector of dates to parse}
}
\value{
parsed datetimes from \link{lubridate}
}
\description{
The date parsing this package will use.
}
\keyword{internal,}
\keyword{validation}
