% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{rename_title_to_cols}
\alias{rename_title_to_cols}
\title{Rename Title columns to internal slot names}
\usage{
rename_title_to_cols(schema, data)
}
\arguments{
\item{schema}{The data schema}

\item{data}{Dataframe, with title slots renamed.}
}
\description{
Schema slots have internal names (characters and underscores only) and "Titles",
which are human readable names and can contain characters that play poorly with internal code.
This function renames title slots to computer-friendly names.
It will not rename slots that are not a part of the standard, and will return them.
}
\keyword{validation}
