% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema_parsing.R
\name{schema_urls}
\alias{schema_urls}
\title{Holds a list of urls from which to download schema files}
\usage{
schema_urls()
}
\description{
Holds a list of urls from which to download schema files
}
\keyword{internal}
