% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{select_cols_of_schema}
\alias{select_cols_of_schema}
\title{Select only those columns that are found in the schema}
\usage{
select_cols_of_schema(schema, data, log = FALSE)
}
\arguments{
\item{schema}{the schema}

\item{data}{dataframe to}

\item{log}{TRUE/FALSE, whether to log the columns removed}
}
\value{
A dataframe, with only the columns from the schema
}
\description{
From a dataframe, select only those columns that are found in the schema.
Note: this function works with both "name" (underscores only) and "title" (human readable).
It also rearranges the columns to match the order of the slots of the schema.
}
\keyword{validation}
