% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema_parsing.R
\name{slot_names}
\alias{slot_names}
\title{Get the slots of a schema}
\usage{
slot_names(schema)
}
\arguments{
\item{schema}{A schema file returned from \link{load_schema}}
}
\description{
Get the slots (or columns) of a schema
}
