% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema_parsing.R
\name{slot_ranges}
\alias{slot_ranges}
\title{Retreive the ranges of a slot}
\usage{
slot_ranges(schema, slot)
}
\arguments{
\item{schema}{A schema file returned from \link{load_schema}}

\item{slot}{A name of a single slot}
}
\value{
A vector of the range names
}
\description{
Returns the 'range' value of a slot.
}
\keyword{internal}
