% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{validate}
\alias{validate}
\title{Validate a dataframe based on the supplied schema.}
\usage{
validate(schema, data, log = c("all", "errors"))
}
\arguments{
\item{schema}{the data schema to use}

\item{data}{a dataframe of the data to test}
}
\value{
Nothing, but prints a log of validation passes and failures.
}
\description{
This function validates the dataframe and prints a log indicating successes and
failures. It first checks for schema columns that are not in the supplied dataframe.
}
\details{
Next, it validates each slot based on the schema's rules.
It first checks to make sure that there
is data in the dataframe of the supplied slot, and that the values are not all NA  #' (otherwise the function exits).
It tests if any NAs are in recommended or required columns.
It checks for duplicates in slots marked as "identifiers".
Finally, it determines the "type" of slot (e.g., date, or menu, string) and validates the data based on the type's respective rules.
}
\keyword{validation}
