% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{validate_rgx}
\alias{validate_rgx}
\title{Validate a string bassed on supplied regex.}
\usage{
validate_rgx(rgx, slot, data, log)
}
\arguments{
\item{rgx}{Regular Expression to test on the data}

\item{data}{A vector of strings to test the expression on}
}
\value{
Nothing, but logs
}
\description{
This function takes a regular expression and tests the supplied data with
it. Since NA's fail validation, remove them, but be explicit about it.
}
\keyword{internal,}
\keyword{validation}
