% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{validate_slot_with_data}
\alias{validate_slot_with_data}
\title{Validate a column of data using slot parameters}
\usage{
validate_slot_with_data(schema, slot, data)
}
\arguments{
\item{schema}{the data schema}

\item{slot}{The name of the slot to test}

\item{data}{A dataframe. The relevant column will be pulled based on the slot name.
The dataframe should have a column with the same name as the slot, but if it doesn't the
function will exit}
}
\value{
Nothing, but prints a log of validation successes and failures.
}
\description{
This is the main validation function.
}
\keyword{internal,}
\keyword{validation}
