% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{validate_values}
\alias{validate_values}
\title{Validate data against permissible values}
\usage{
validate_values(schema, slot, x)
}
\value{
Logical vector
}
\description{
Note: This will return TRUE if the value is NA! We should catch this
seperately.
}
\keyword{internal,}
\keyword{validation}
