% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema_parsing.R
\name{get_menu_values}
\alias{get_menu_values}
\title{Get the permissible values of a range, if they exist}
\usage{
get_menu_values(schema, ranges)
}
\arguments{
\item{schema}{A schema file returned from \link{load_schema}}

\item{ranges}{A list of ranges, returned from \link{slot_ranges}}
}
\value{
A vector of permissible values for each range, NULL if none exist
}
\description{
All slots have ranges. Some of these ranges have values
Get the permissible values of each range,
}
\keyword{internal,}
\keyword{parsing}
