% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{log_basic_validation}
\alias{log_basic_validation}
\title{Print out if validation passed or failed}
\usage{
log_basic_validation(x, slot, data, pass_message, fail_message, ids = NULL)
}
\arguments{
\item{x}{Logical vector indication pass/fail of validation}

\item{slot}{slot name}

\item{data}{the vector of data that was validated}

\item{pass_message}{Message to send to \link{log_pass}}

\item{fail_message}{Message to send to \link{log_error}}

\item{ids}{(optional) vector of the ids, for printing.}
}
\value{
Nothing, but prints out if validation succeeded or failed.
}
\description{
Print out if validation passed or failed
}
\keyword{internal,}
\keyword{logging,}
\keyword{validation}
